feat(inventory): align quantity types and add sell stock and integrate with order creation

+++ internal/services/products/model.go
@@ -36,5 +36,5 @@ type ProductListItem struct {
 	Price       int64     `bun:"price"`
 	CategoryID  uuid.UUID `bun:"category_id"`
 	ImageURL    *string   `bun:"image_url"`
-	Quantity    int64     `bun:"quantity"`
+	Quantity    int       `bun:"quantity"`
 }
